+++ rust/src/pr006.rs
@@ -19,10 +19,10 @@
 
 define_problem!(pr006, 6, 25164150);
 
-fn pr006() -> usize {
+fn pr006() -> u64 {
     let mut sum_sq = 0;
     let mut sum = 0;
-    for i in 1u .. 101 {
+    for i in 1u64 .. 101 {
         sum += i;
         sum_sq += i * i;
     }
